requirements-dev.txt
-r requirements.txt

# Dev
black==21.7b0
python-dotenv==0.19.0

# Linter
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6

# Tests
pytest==6.2.4
pytest-flask==1.2.0
pytest-dotenv==0.5.2
mock==4.0.3

# Devs
flask-shell-ipython==0.4.1
